# Speed up `unnest_longer()` and `unnest_wider()`