--- a/R/rectangle.R
+++ b/R/rectangle.R
@@ -402,7 +402,7 @@ simplify_col <- function(x, nm, ptype = list(), transform = list(), simplify = F
     n <- map_int(x, vec_size)
   }
 
-  if (!all(n %in% c(0, 1))) {
+  if (any(n > 1)) {
     if (is.null(ptype)) {
       return(x)
     } else {