Changes
* replace `map()` and `compact()` with `lapply()` and `vec_slice() + list_sizes()`
* replace `tibble()` with `new_data_frame()`

They should also benefit nicely from #1126.

## Benchmarks

``` r
library(dplyr, warn.conflicts = FALSE)
library(purrr)
library(tidyr)

array_fields <- c("titles", "aliases", "allegiances", "books", 
                  "povBooks", "tvSeries", "playedBy")
got_chars2 <- map(repurrrsive::got_chars, 
                  ~map_at(.x, array_fields, ~as.list(.x)))

times <- tibble(
  method = character(),
  rows = integer(),
  time = numeric()
)

data <- tibble(data = rep(got_chars2, length.out = 5000))

bench::mark(
  unnest_longer = unnest_longer(data, data),
  unnest_wider = unnest_wider(data, data, simplify = TRUE),
  check = FALSE
)

# This PR
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_longer    192ms    199ms      4.98    5.55MB     16.6
#> 2 unnest_wider     911ms    911ms      1.10    14.6MB     24.1

# CRAN
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_longer    2.33s    2.33s     0.430      12MB     15.5
#> 2 unnest_wider     2.33s    2.33s     0.430    17.4MB     23.2
```

<sup>Created on 2021-06-17 by the [reprex package](https://reprex.tidyverse.org) (v2.0.0)</sup>